calsipro/pyproject.toml
[tool.ruff]
line-length = 120

[tool.poetry]
name = "calsipro"
version = "0.16.0"
description = ""
authors = ["Simon Haendeler <simon.ac@haend.de>"]

[tool.poetry.scripts]
calsipro = 'calsipro.cli:cli'

[tool.poetry.dependencies]
python = "^3.8"
openpyxl = "^3.0.10"
datashader = "^0.14.2"
numba = [{version = "^0.56.0", python = ">=3.8,<3.11"}, {version = "^0.57", python=">=3.11"}]
Pillow = "^9.2.0"
click = "^8.1.3"
numpy = ">=1.18"
scipy = "^1.9.0"
matplotlib = "^3.5.3"
bokeh = "^3.0.0"
aicsimageio = "^4.9.4"
aicspylibczi = ">=3.0.5"
fsspec = ">=2022.7.1"
syn-bokeh-helpers = {version = "^0.5.0", source = "syntonym"}
pywavelets = "^1.4.1"
ffmpeg-python = "^0.2.0"
scikit-learn = "^1.2.1"
polars = "^0.16.14"
s3fs = "^2023.5.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
mypy = "^0.991"
popy = {version = "^0.1.1", source = "syntonym"}


[[tool.poetry.source]]
name = "syntonym"
url = "http://localhost:8080/"
default = false
secondary = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
